nlp-fine-tuning-api/requirements.txt
transformers>=4.30.0
datasets>=2.12.0
torch>=1.13.0
scikit-learn
pandas
mlflow
fastapi
uvicorn[standard]
pydantic
pyyaml
sentencepiece